+++ hack/kind-cluster.sh
@@ -18,15 +18,41 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+set -e
 
-set -o errexit
+install_kind() {
+  echo "Installing KIND cluster"
+  GO111MODULE="on" go get sigs.k8s.io/kind@v0.9.0
+}
 
-set -o pipefail
+create_kind_cluster() {
+  install_kind
+  echo "creating KIND cluster"
+  kind create cluster --name kind-cicd
+}
 
-export CONFIG_PATH="$(pwd)/test"
+destroy_kind_cluster() {
+  echo "destroying KIND cluster"
+  kind delete cluster --name kind-cicd
+}
 
-# Run unit and integration tests excluding dependencies
-PACKAGES=$(go list ./... | grep -v '/vendor/')
-for package in $PACKAGES; do
-    go test -tags=test ${@} "$package" -v
-done
+help() {
+  usage="$(basename "$0") [option] -- Script to create or destroy KIND cluster.
+  Available options are destroy-kind, create-kind or help"
+  echo $usage
+}
+
+
+if [ $# -gt 1 ]; then help ;fi
+case "${1}" in
+  create-kind)
+    create_kind_cluster
+    ;;
+  destroy-kind)
+    destroy_kind_cluster
+    ;;
+  *)
+    help
+    exit 1
+esac
+