+++ pkg/bot/slack.go
@@ -127,8 +127,6 @@ func (b *SlackBot) Start() {
 }
 
 func (sm *slackMessage) HandleMessage(b *SlackBot) {
-	log.Debugf("Slack incoming message: %+v", sm.Event)
-
 	// Check if message posted in authenticated channel
 	info, err := sm.SlackClient.GetConversationInfo(sm.Event.Channel, true)
 	if err == nil {
@@ -159,6 +157,8 @@ func (sm *slackMessage) HandleMessage(b *SlackBot) {
 }
 
 func (sm *slackMessage) Send() {
+	log.Debugf("Slack incoming Request: %s", sm.Request)
+	log.Debugf("Slack Response: %s", sm.Response)
 	// Upload message as a file if too long
 	if len(sm.Response) >= 3990 {
 		params := slack.FileUploadParameters{
@@ -177,7 +177,14 @@ func (sm *slackMessage) Send() {
 		return
 	}
 
-	if _, _, err := sm.RTM.PostMessage(sm.Event.Channel, slack.MsgOptionText("```"+sm.Response+"```", false), slack.MsgOptionAsUser(true)); err != nil {
+	var options = []slack.MsgOption{slack.MsgOptionText("```"+sm.Response+"```", false), slack.MsgOptionAsUser(true)}
+
+	//if the message is from thread then add an option to return the response to the thread
+	if sm.Event.ThreadTimestamp != "" {
+		options = append(options, slack.MsgOptionTS(sm.Event.ThreadTimestamp))
+	}
+
+	if _, _, err := sm.RTM.PostMessage(sm.Event.Channel, options...); err != nil {
 		log.Error("Error in sending message:", err)
 	}
 }