+++ test/e2e/e2e_test.go
@@ -34,6 +34,9 @@ import (
 	"github.com/infracloudio/botkube/test/e2e/env"
 	"github.com/infracloudio/botkube/test/e2e/filters"
 	"github.com/infracloudio/botkube/test/e2e/notifier/create"
+	"github.com/infracloudio/botkube/test/e2e/notifier/delete"
+	"github.com/infracloudio/botkube/test/e2e/notifier/error"
+	"github.com/infracloudio/botkube/test/e2e/notifier/update"
 	"github.com/infracloudio/botkube/test/e2e/welcome"
 )
 
@@ -84,6 +87,9 @@ func TestRun(t *testing.T) {
 		"notifier": create.E2ETests(testEnv),
 		"command":  command.E2ETests(testEnv),
 		"filters":  filters.E2ETests(testEnv),
+		"update":   update.E2ETests(testEnv),
+		"delete":   delete.E2ETests(testEnv),
+		"error":    error.E2ETests(testEnv),
 	}
 
 	// Run test suite