+++ test/e2e/notifier/create/create.go
@@ -21,7 +21,6 @@ package create
 
 import (
 	"encoding/json"
-	"fmt"
 	"testing"
 
 	"github.com/nlopes/slack"
@@ -129,18 +128,8 @@ func (c *context) testCreateResource(t *testing.T) {
 				assert.Equal(t, test.ExpectedWebhookPayload.Summary, lastSeenPayload.Summary)
 			}
 
-			resource := fmt.Sprintf("%s/%s/%s", test.GVR.Group, test.GVR.Version, test.GVR.Resource)
-			if test.GVR.Group == "" {
-				resource = fmt.Sprintf("%s/%s", test.GVR.Version, test.GVR.Resource)
-			}
-			isAllowed := utils.AllowedEventKindsMap[utils.EventKind{
-				Resource:  resource,
-				Namespace: "all",
-				EventType: config.CreateEvent}] ||
-				utils.AllowedEventKindsMap[utils.EventKind{
-					Resource:  resource,
-					Namespace: test.Namespace,
-					EventType: config.CreateEvent}]
+			resource := utils.GVRToString(test.GVR)
+			isAllowed := utils.CheckOperationAllowed(utils.AllowedEventKindsMap, test.Namespace, resource, config.CreateEvent)
 			assert.Equal(t, isAllowed, true)
 		})
 	}